requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Deprecated==1.2.10
htmlmin==0.1.12
idna==2.9
Mako==1.1.2
MarkupSafe==1.1.1
PyGithub==1.51
PyJWT==1.7.1
requests==2.23.0
urllib3==1.25.9
wrapt==1.12.1
